Building_BankEase-main/bankease_agentic_app/requirements.txt
# Notice that for faiss-cpu below, a particular version is used. This was necessary
# because when no version was used, was getting errors of type " ERROR: Failed building wheel for faiss-cpu"
# on Mac OS platform.
langchain 
langchain_core
langchain_community
langchain_groq
langchainhub
langgraph
langchain_huggingface
faiss-cpu==1.7.4
pandas
sentence-transformers
streamlit
seaborn
matplotlib
squarify

# Following added for classic ML based recommendation model
fastapi
uvicorn[standard]
numpy
scikit-learn
joblib
lightgbm
scipy
scikit-surprise
pydantic
